requirements.txt
requests
pyexiftool
tdqm